requirements.txt
os
glob
typing
python-dotenv
multiprocessing
tqdm
langchain
chromadb
fastapi
pydantic
argparse
cohere
llama-cpp-python
sentence_transformers
urllib3
PyMuPDF
unstructured
extract-msg
tabulate
pandoc
pypandoc
